requirements.txt
langchain
langchain_community
langchain-google-genai
langchain_experimental
python-dotenv
streamlit
langchain_experimental
mysql-connector-python
pymysql

